third_party/blink/web_tests/http/tests/lcp-script-observer/lcp-img-creation-vs-insertion.html

<!doctype html>
<html>

<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>

<body />
<!-- inert script is loaded but does not contribute to LCP -->
<script src="/lcp-script-observer/resources/inert.js"></script>
<script src="/lcp-script-observer/resources/lcp-img-create.js" async></script>
<script src="/lcp-script-observer/resources/lcp-img-insert.js" async></script>
<script>
  assert_implements(window.LargestContentfulPaint, "LargestContentfulPaint is not implemented");

  promise_test(() => {
    return new Promise((resolve, reject) => {
      const observer = new PerformanceObserver(list => {
        const entries = list.getEntries();
        // find LCP image candidate
        const lcp = entries.find(entry => entry.element.nodeName === 'IMG');
        if (!lcp) {
          reject();
        }

        // ensure we received a valid LCP candidate
        assert_true(Boolean(lcp));

        // ensure LCP element is an image.
        assert_equals(lcp.element.nodeName, 'IMG');

        // assert LCP image knows its creator scripts.
        const creators = internals.getCreatorScripts(lcp.element)
          .map(url => new URL(url).pathname);
        assert_array_equals(creators, [
          '/lcp-script-observer/resources/lcp-img-insert.js',
          '/lcp-script-observer/resources/lcp-img-create.js',
        ]);
        resolve(lcp);
      });

      observer.observe({type: 'largest-contentful-paint', buffered: true});
    });
  }, 'LCP element knows its dependencies, including creation and insertion scripts');

</script>

</html>
